style: format the tree and fix all clippy findings
cargo fmt --check reported 110 hunks in 38 files and cargo clippy failed
on the default build with two approx_constant errors and 38 warnings.

- Use core::f32::consts::PI instead of hand-written constants.
- Replace the duplicated solve_linear_9 / solve_linear_3 Gauss-Jordan
  solvers with one const-generic solve_linear<N>.
- Replace the 8-argument imu::channels::publish with a Publisher that
  owns the revision counter; MagneticState::report bundles the
  magnetometer health fields.
- Pass line endpoints as Point / tuples instead of four scalars.
- Mechanical fixes: is_multiple_of, div_ceil, let chains, redundant
  casts, needless range loops, ok().expect() on a non-Debug error.

+++ src/ui/common.rs
@@ -50,7 +50,7 @@ fn raw_color(raw: u16) -> Rgb565 {
 }
 
 pub(crate) fn fill_rect(frame: &mut GuiFramebuffer, rect: Rect, color: Rgb565) {
-    fill_box(frame, rect.x, rect.y, rect.w as u32, rect.h as u32, color);
+    fill_box(frame, rect.x, rect.y, rect.w, rect.h, color);
 }
 
 pub(crate) fn fill_box(
